update remove City and Site

+++ projetSortir/src/Controller/CityController.php
@@ -4,9 +4,13 @@ namespace App\Controller;
 
 use App\Entity\City;
 use App\Form\CityType;
+use App\Repository\CityRepository;
+use App\Repository\UserRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 
@@ -42,7 +46,7 @@ class CityController extends AbstractController
 
             $this->addFlash("success", "Ville modifié"); // info warning error
 
-            return $this->redirectToRoute('home');
+            return $this->redirectToRoute('city');
         }
 
         return $this->render('city/edit.html.twig', [
@@ -69,7 +73,7 @@ class CityController extends AbstractController
 
             $this->addFlash("success", "Ville modifié"); // info warning error
 
-            return $this->redirectToRoute('home');
+            return $this->redirectToRoute('city');
         }
 
         return $this->render('city/edit.html.twig', [
@@ -79,19 +83,24 @@ class CityController extends AbstractController
     }
 
     /**
-     * @Route("/deleteCity", name="deleteCity")
+     * @Route("/removeCity", name="removeCity")
      * @IsGranted("ROLE_ADMIN")
      */
-    public function deleteCity(Request $request, EntityManagerInterface $em)
+    public function removeCity(Request $request , EntityManagerInterface $em , CityRepository $cityRepository)
     {
-        $citiesRepository = $em->getRepository(City::class);
-        $city = $citiesRepository->find($request->get('cityId'));
+        $response = new JsonResponse();
+        $city = $cityRepository->find($request->request->get('id'));
 
-        $em->remove($city);
-        $em->flush();
-
-        $this->addFlash("success", "Ville supprimée"); // info warning error
-
-        return $this->redirectToRoute('home');
+        if (!$city) {
+            $response->setContent(json_encode(['msg' => 'Ville introuvable']));
+            $response->setStatusCode(Response::HTTP_BAD_REQUEST);
+        } else {
+            $em->remove($city);
+            $em->flush();
+            $response->setContent(json_encode(['msg' => "Suppression de la ville réussit" ]));
+        }
+        return $response;
     }
+
+
 }
